feat(msvc): enable MSVC by compiling C tests as C++ - remove hard MSVC ban - route MSVC builds through C++ (/TP, /EHsc) - refactor TEC_TRY_BLOCK for C and C++ - minor warnings fix for MSVC in tests/ - add registration tests for macro-generated symbols

+++ include/string_utils.h
@@ -1,9 +1,10 @@
 #ifndef STRING_UTILS_H
 #define STRING_UTILS_H
 
-char* string_concat(const char* a, const char* b);
-int string_length(const char* str);
-char* string_reverse(const char* str);
-char* string_upper(const char* str);
+#include <stddef.h>
+char *string_concat(const char *a, const char *b);
+size_t string_length(const char *str);
+char *string_reverse(const char *str);
+char *string_upper(const char *str);
 
 #endif
